Bugfix: Segmha importer plugin without inputs makes impossible to associate segmentations with a stack.
- Analysis check on file load now optional. Modified application settings and general settings panel.
- Fix: Noise in xml buffer when writing extensions in Channel::snapshot().
- Temporal storage now throws exceptions on failure to create directories or saving snapshots.
- Remove unused code in SourceFilter.
- Fix: Segmha importer and source filters without inputs make impossible to associate stacks with segmentations.
- Fix: invalid filename for windows OS in tabular report.

+++ Core/Analysis/Filter.cpp
@@ -176,7 +176,12 @@ void Filter::restoreEditedRegions()
 {
   if (validStoredInformation())
   {
-    QByteArray buffer = storage()->snapshot(outputFile());
+    QByteArray buffer;
+
+    if(storage()->exists(outputFile()))
+    {
+      buffer = storage()->snapshot(outputFile());
+    }
 
     if (!buffer.isEmpty())
     {
@@ -243,13 +248,14 @@ bool Filter::existOutput(Output::Id id) const
 //----------------------------------------------------------------------------
 bool Filter::restorePreviousOutputs() const
 {
-//   qDebug() << "Restore Previous Outputs Request: " << m_type << uuid();
   if (validStoredInformation())
   {
-//     qDebug() << " - Accepted - outputfile" << outputFile();
-    QByteArray buffer = storage()->snapshot(outputFile());
+    QByteArray buffer;
 
-//     qDebug() << buffer;
+    if(storage()->exists(outputFile()))
+    {
+      buffer = storage()->snapshot(outputFile());
+    }
 
     if (!buffer.isEmpty())
     {
